[VC] Split GsCodeLanguageServer into partial files
Refactor: extract LSP handlers and helpers from GsCodeLanguageServer into multiple partial files to improve organization and readability. Added new files: Completion, DocumentSymbol, DocumentSync, Helpers, Lifecycle, Navigation, SemanticTokens, and Workspace under server/GSCode.NET/LSP. Updated GsCodeLanguageServer.cs to be a partial class and removed the in-file handler implementations (note at EOF indicates remainder moved to partial files). Functionality and LSP capabilities are preserved; this is a code-structure/maintenance-only change.
